requirements.txt
flask
psycopg2
python-dotenv
asyncpg
pyserial